refactor DHS one-sided code
