+++ autode/bracket/imagepair.py
@@ -358,42 +358,6 @@ class BaseImagePair(ABC):
     def dist(self):
         """Distance defined between two images in the image-pair"""
 
-    def get_coord_by_side(self, side: str) -> OptCoordinates:
-        """For external usage, supplies only the coordinate object"""
-        _, coord, _, _ = self._get_img_by_side(side)
-        # todo put these functions in DHS if not needed in base
-        return coord
-
-    def _get_img_by_side(
-        self, side: str
-    ) -> Tuple["Species", OptCoordinates, _OptimiserHistory, float]:
-        """
-        Access an image and some properties by a string that
-        represents side. Returns a tuple of the species, the
-        current coordinate object, and a factor that is necessary
-        for some calculations
-
-        Args:
-            side (str): 'left' or 'right'
-
-        Returns:
-            (tuple) : tuple(image, current coord, history, fac)
-        """
-        if side == "left":
-            img = self._left_image
-            coord = self.left_coord
-            hist = self._left_history
-            fac = 1.0
-        elif side == "right":
-            img = self._right_image
-            coord = self.right_coord
-            hist = self._right_history
-            fac = -1.0
-        else:
-            raise ImgPairSideError()
-
-        return img, coord, hist, fac
-
     @property
     def _total_history(self) -> _OptimiserHistory:
         """
@@ -583,107 +547,3 @@ class ImagePair(BaseImagePair, ABC):
             (Distance): Distance in Angstrom
         """
         return Distance(np.linalg.norm(self.dist_vec), units="ang")
-
-    def update_one_img_mol_energy(self, side: str) -> None:
-        """
-        Update only the molecular energy using the supplied
-        engrad_method for one image only
-
-        Args:
-            side (str): 'left' or 'right'
-        """
-        assert self._engrad_method is not None
-        assert self._n_cores is not None
-        img, coord, _, _ = self._get_img_by_side(side)
-
-        logger.debug(
-            f"Calculating energy for {side} side"
-            f" with {self._engrad_method}"
-        )
-
-        en = _calculate_energy_for_species(
-            species=img.copy(),
-            method=self._engrad_method,
-            n_cores=self._n_cores,
-        )
-        # update coord
-        coord.e = en.to("Ha")
-
-    def update_one_img_mol_engrad(self, side: str) -> None:
-        """
-        Update the molecular energy and gradient using the supplied
-        engrad_method for one image only
-
-        Args:
-            side (str): 'left' or 'right'
-        """
-        assert self._engrad_method is not None
-        assert self._n_cores is not None
-        img, coord, _, _ = self._get_img_by_side(side)
-
-        logger.debug(
-            f"Calculating engrad for {side} side"
-            f" with {self._engrad_method}"
-        )
-        en, grad = _calculate_engrad_for_species(
-            species=img.copy(),
-            method=self._engrad_method,
-            n_cores=self._n_cores,
-        )
-        # update coord
-        coord.e = en.to("Ha")
-        coord.update_g_from_cart_g(grad.to("Ha/ang"))
-        return None
-
-    def update_one_img_mol_hess_by_calc(self, side: str) -> None:
-        """
-        Updates the molecular hessian using supplied hess_method
-        for one image only
-
-        Args:
-            side (str): 'left' or 'right'
-        """
-        assert self._hess_method is not None
-        assert self._n_cores is not None
-        img, coord, _, _ = self._get_img_by_side(side)
-
-        logger.debug(
-            f"Calculating Hessian for {side} side" f" with {self._hess_method}"
-        )
-        hess = _calculate_hessian_for_species(
-            species=img.copy(), method=self._hess_method, n_cores=self._n_cores
-        )
-        # update coord
-        coord.update_h_from_cart_h(hess.to("Ha/ang^2"))
-        return None
-
-    def update_one_img_mol_hess_by_formula(self, side: str) -> None:
-        """
-        Updates the molecular hessian of one side by using Hessian
-        update formula; requires the gradient and hessian for the
-        last coordinates, and gradient for the current coordinates
-
-        Args:
-            side (str): 'left' or 'right'
-        """
-        img, coord, hist, _ = self._get_img_by_side(side)
-        assert len(hist) > 1, "Hessian update not possible!"
-        assert coord.h is None, "Hessian already exists!"
-        assert coord.g is not None, "Gradient should be present!"
-        last_coord = hist.penultimate
-        for update_type in self._hessian_update_types:
-            updater = update_type(
-                h=last_coord.h,
-                s=coord.raw - last_coord.raw,
-                y=coord.g - last_coord.g,
-                subspace_idxs=coord.indexes,
-            )
-            if not updater.conditions_met:
-                continue
-
-            coord.h = updater.updated_h
-            break
-
-        assert coord.h is not None, "Hessian update failed!"
-
-        return None
